// src/app/Shared/Models/OrderDetails.ts
export class OrderDetails {
  subJectArea: string;
  typeOfDocument: string;
  numberOfWords: string;
  spacing: any;
  days: number;
  hours: number;
  dueDate: any;
  rewriter: any;
  academicLevel: string;
  noOfReferences: number;
  style: string;
  currency: string;
  note: string;
  preferredLanguage: string;
  clientId: number;
  urgency: string;
  discipline: any;
}
export class OrderViewModel {
  SubJectArea: string;
  TypeOfDocument: string;
  NumberOfWords: string;
  Spacing: any;
  Urgency: string;
  DueDate: any;
  Rewriter: any;
  AcademicLevel: string;
  NoOfReferences: number;
  Style: string;
  Currency: string;
  Note: string;
  preferredLanguage: string;
  ClientId: number;
  OrderNumber: string;
  OrderDate: any;
  Title: string;
  OrderCost: string;
  Paid: string;
  CompletedDate: any;
  Spaces: string;
  Status: string;
  OrderValue: number;
}